+++ core/app/models/spree/order.rb
@@ -509,10 +509,13 @@ module Spree
     end
 
     # Tells us if there if the specified promotion is already associated with the order
-    # regardless of whether or not its currently eligible.  Useful because generally
-    # you would only want a promotion to apply to order no more than once.
-    def promotion_credit_exists?(promotion)
-      !! adjustments.promotion.reload.detect { |credit| credit.originator.promotion.id == promotion.id }
+    # regardless of whether or not its currently eligible. Useful because generally
+    # you would only want a promotion action to apply to order no more than once.
+    #
+    # Receives an adjustment +originator+ (here a PromotionAction object) and tells
+    # if the order has adjustments from that already
+    def promotion_credit_exists?(originator)
+      !! adjustments.promotion.reload.detect { |credit| credit.originator.id == originator.id }
     end
 
     def promo_total